refactor: move yamlfmt to separate module
